+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 <!-- ANCHOR: changelog -->
 
+### Added
+
+- Add `default` keyword arg to `env()`
+
 ## [4.2.1] - 2025-11-26
 
 <!-- ANCHOR: changelog -->